HW4.1 correction

+++ src/main/java/ru/hogwarts/school/controllers/StudentController.java
@@ -84,7 +84,7 @@ public class StudentController {
     }
 
     @GetMapping("/middle-age-students")
-    public ResponseEntity<Integer> getMiddleAgeOfStudents() {
+    public ResponseEntity<Double> getMiddleAgeOfStudents() {
         return ResponseEntity.ok(studentService.getMiddleAgeOfStudent());
     }
     @GetMapping("/five-last-student")
